Networking: Introducing external-dns
This major change was created by the replacement of a lot of logic with
a dependency called external-dns(https://github.com/kubernetes-sigs/external-dns). The benefits of using external dns is massive when it comes to dealing with cloud providers as it already has most of the tools needed for dealing with them. There are a few shortcomings that I'm still working on, but there seems to be possible solutions to them. Also, for the user, this should make it much easier to them to try out sequencer as external-dns and cert-manager are already tools that most devops team are familiar with.

The documentation has been completely revamped to include external-dns.

Also moved providers in the code to a more specific `tunneling` package
as that's the only thing really that it does.

Hopefully, the codebase is easier to navigate as a result of those
changes