DNS: Add DNS Condition to better support external-dns
The first implementation wasn't going to work with ingresses and routes,
these changes reintroduce the concept of DNS condition to Workspaces so
that DNS can be handled in isolation. Workspace also was added a new
status field for DNS section so it can reconcile external's endpoints
with what is defined by the user through Workspace.