Dev: Update all Kustomize specs to work with different namespaces
The current solver needs access to `kube-system` and as such, the whole
set of kustomize templates needed to be reworked so it can deal with
different namespace. I've always wanted to streamline the kustomize
templates anyway as those are only going to be used to generate
templates from kubebuilder comment annotation in the code and for
development purposes.

I'd also like to rework the cert-manager integration there, but this
change is already massive as it is.