DNS: Adding new core type & ingress info
First the ingress load balancer service is added to the spec to make it
easier in this version to create the network pipe, it's possible in the
future it can be removed or at least made optional, but right now,
the complexity of inferring the loadbalancer service is problematic and
could cause more confusion to the user. Documentation will need to be
added to add information on how to find the load balancer service.

Second, I'm adding a new core type called DNSRecord. This record is
going to replace external-dns entirely and will be the entrypoint where
any record needs to be created. Annotations are absolutely useful for
permanent system, but the way DNS records are used within Sequencer
makes it hard to fit within the external-dns structure (even with the
CRDs). For that reason, a new CRD needs to be created to manage this
part of the stack.