DNS: Reorganize & add fields to status
Created a dnsrecords package to hold all the types that are needed for
DNSRecord, also added some more fields to the status to make it possible
to store provider-specific information. The status is viewed as
experimnetal and will most likely change in the near future as more
providers are added.