DNS: Workspace DNS Association rework
First renamed Settings->Properties as part of the DNSRecord spec as that
feels more natural after playing with both options.

Then, removed the DNS status of the Workspace subresource as now the
logic is going to be exclusively handled by the new DNSRecord Custom
Resource.

Those DNSRecord will still be fetchable from a Workspace as the
OwnerReference will be assigned.