Workspaces: Rework ingress condition flow
The Ingress condition flow could be stuck in an infinite check because
of the way it returned some marker. Refactored the way the condition is
handled so the DNS records are only set once in the Status for the DNS
reconciliation task to be not overflow.

The handling of the waiting condition was also reworked to be a little
bit simpler. It's possible the Guard concept added to the provider
Interface needs to be applied to condition handling in general. For now,
the code behaves in a similar way but doesn't follow the same structure
exactly.